pyproject.toml
[project]
name = "agentic-saul"
version = "0.1.0"
description = "A legal research assistant powered by LangGraph and Chainlit"
requires-python = ">=3.9,<3.12"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "chainlit~=2.0.4",
    "langgraph~=0.2.67",
    "langchain~=0.3.15",
    "langchain-community~=0.3.16",
    "langchain-openai~=0.3.2",
    "langchain-qdrant~=0.2.0",
    "langchain_huggingface~=0.1.2",
    "wikipedia~=1.4.0",
    "praw~=7.8.1",
    "semanticscholar~=0.9.0",
    "python-dotenv~=1.0.1",
    "websockets~=14.2",
    "google-search-results~=2.4.2",
    "loguru~=0.7.3",
    "ipykernel~=6.29.5",
    "ipywidgets~=8.1.5",
    "pypdf~=5.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff~=0.3.3",
    "black~=24.2.0",
    "mypy~=1.9.0",
]

[tool.setuptools]
packages = ["saul"]

[tool.ruff]
select = ["E", "F", "I", "N", "W", "B"]
line-length = 100

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true 